添加接口管理

+++ src/main/java/org/lsqt/components/mvc/util/ActionFormUtil.java
@@ -3,6 +3,8 @@ package org.lsqt.components.mvc.util;
 import java.io.File;
 import java.lang.reflect.Field;
 import java.lang.reflect.Parameter;
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.Collection;
@@ -239,15 +241,15 @@ public final class ActionFormUtil {
 		} else if (baseType == byte.class) {
 			return value == null ? (byte) 0 : Byte.valueOf(value.toString());
 		}
-		/*
+		
 		else if(BigInteger.class.isAssignableFrom(baseType)) {
 			return value == null ? null:BigInteger.valueOf(Long.valueOf(value.toString()));
 		}
 		
 		else if(BigDecimal.class.isAssignableFrom(baseType)) {
-			return value == null ? null:BigDecimal.valueOf(Long.valueOf(value.toString()));
+			return value == null ? null:BigDecimal.valueOf(Double.valueOf(value.toString()));
 		}
-		*/
+		
 		
 		
 		else if (Boolean.class.isAssignableFrom(baseType)) { // 布尔型
